chore: Bump nvidia-ml-py to v12.555.43
The datadog agent upgraded to Python 3.11 in v7.51.0 and at the same time
uprgaded `google/protobuf` to a version that is unsupported by the very
old version of `nvidia-ml-py`.

This upgrades `nvidia-ml-py` to the latest version. Some APIs related to
throttling was no longer in the library, so I removed the reference to
those.

This is confirmed working with NVML v555.42.02.